Déplace les écrans des nouvelles pages
