+++ assets/scripts/routes/create-new-site.js
@@ -3,7 +3,7 @@ import page from 'page'
 import { svelteTarget } from "../config";
 import { replaceComponent } from "../routeComponentLifeCycle";
 import store from "../store";
-import CreateNewSite from '../components/CreateNewSite.svelte'
+import CreateNewSite from '../components/screens/CreateNewSite.svelte'
 import { fetchAuthenticatedUserLogin } from '../actions.js'
 export default () => {
     fetchAuthenticatedUserLogin()
