+++ assets/scripts/routes/select-or-create-site.js
@@ -3,7 +3,7 @@ import page from 'page'
 import { svelteTarget } from "../config";
 import { replaceComponent } from "../routeComponentLifeCycle";
 import store from "../store";
-import SelectOrCreateSite from '../components/SelectOrCreateSite.svelte'
+import SelectOrCreateSite from '../components/screens/SelectOrCreateSite.svelte'
 
 export default () => {
     const selectOrCreateSite = new SelectOrCreateSite({
